requirements.txt
pandas
numpy
python-binance
scikit-learn
joblib
streamlit
plotly
streamlit-autorefresh
python-dotenv
